message-api-crossfitlov/requirements.txt
Flask==1.1.2
Flask-BasicAuth==0.2.0
mysql-connector-python==8.0.23
requests==2.25.1